Remove persitence cache for known instances
