--- src/ServiceControl.Audit/Monitoring/KnownEndpointIndex.cs
@@ -1,20 +1,0 @@
-namespace ServiceControl.CompositeViews.Endpoints
-{
-    using System.Linq;
-    using EndpointControl;
-    using Raven.Client.Indexes;
-
-    class KnownEndpointIndex : AbstractIndexCreationTask<KnownEndpoint>
-    {
-        public KnownEndpointIndex()
-        {
-            Map = messages => from message in messages
-                select new
-                {
-                   message.Id
-                };
-
-            DisableInMemoryIndexing = true;
-        }
-    }
-}
